gck.env: Export all changed environment variables
Previously we would try to parse `gck.env` in a container.
The better way is to export all Makefile exported variables
that were changed during make execution already in a bash format.

This allows for easy inclusion and potentially conditional syntax
in `gck.env` (not desired).

This also makes the `gck.env` to only be of `make` syntax instead
of `bash` and `make`.